backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.2
python-multipart==0.0.6
pillow==10.1.0
requests==2.31.0
azure-cognitiveservices-vision-customvision==3.0.0
msrest==0.7.1
python-dotenv==1.0.0
flask==2.3.3
